Get item display names from extra data if possible.
This mostly fixes #90 which complains about not using the item's
custom name if it has one. It's not an all-the-way fix because
we don't currently notice if an item is enchanted after we have
cached a HUD item for it, so we don't update the name. I love
cache invalidation problems.