Get a custom name from extra data for real...
I think the implementation of GetDisplayName() on extra data
lists in CommonLibSSE is buggy or maybe outright wrong. This
implementation looks for extra data with the type kTextDisplayData
and interrogates that for display names. The clib implementation
does not filter for extra text data, IIUC.

Anyway. This really fixes for real bug #90 in all cases.