Fetch and display charge for enchanted items
We query enchantment item for the max, and extra data for current
level. This is what the method on InventoryEntryData does, but we
do it without crashing. I think it's GetType<class>() failing where
GetByType(enum) succeeds.

Then we draw a version of the progress bar clipped to show the
filled percentage. Colors are bad, but it works!

Also ran a format.