Reworked the layout data structures again
We now support drawing a fill image over the bg image for the
progress bar, as well as the previous clipped version of the
background image. The concept of vertical/horizontal is gone,
replaced with a single angle that is used to rotate the whole
progress bar. This gives people some nice options.