Unify rust & C++ logs using the Rust facilities.
The choice was to write macros to replace the log crate macros, or
write templates to sink the C++ logging to Rust. I went with Rust
because what good is being in the evangelism task force if I don't
ruthlessly carcinicize?

Wrote the templates by contemplating what the SKSE peeps did to put
templates in front of `spdlog`. These templates are in front of `fmt`
instead. The guts of each implementation routes to shims in front of
the the `log` crate's macros.

The new namespace is `rlog`. All calls to `logger::info()` are now
calls to `rlog::info()` which eventually dispatches to `log::info!()`.

Demoted some logs while I was looking at them in the unified context.
I will need to systematically review logging because it feels too
voluminous now that it's all in one file.