Unify rust & C++ logs using the Rust facilities. (#92)
The choice was to write macros to replace the log crate macros, or
write templates to sink the C++ logging to Rust. I went with Rust
because what good is being in the evangelism task force if I don't
ruthlessly carcinicize?

Wrote the templates by contemplating what the SKSE peeps did to put
templates in front of `spdlog`. These templates are in front of `fmt`
instead. The guts of each implementation routes to shims in front of
the the `log` crate's macros.

The new namespace is `rlog`. All calls to `logger::info()` are now
calls to `rlog::info()` which eventually dispatches to `log::info!()`.

The debug and trace level log templates add filename & line number
information to the start of each log message. Unfortunately those two
levels are decorated with fairly useless information from the rust
logger, which I can't rid of without also affecting rust-sourced log
lines. I'll need to find a new crate. This is, however, good enough 
to ship.

Got rid of spdlog in the build. It remains in the deps because the
CommonLibSSE dll needs it.

Demoted some logs while I was looking at them in the unified context.
I will need to systematically review logging because it feels too
voluminous now that it's all in one file.