Added a few ammo icons.
Made the color keywords function return an option so we can
easily specify a color if none is provided in keywords. Did
this for fire arrows.

Renamed ammo icons to fit the most general -> most specific
qualifier sequence programmers and the military love so much.

Finally, skip the soulsy icon set completeness check in CI,
because we know for sure it fails right now because of missing
ammo icons. I haven't yet sourced good icons for arrowhead types.