Consternation uproar! Reorganize the repo.
The assets and other non-compiled files in the repo are now
arranged in a directory named `installer` to make creating a
FOMOD installer extremely simple. The archive-building scripts
and the installation script can now get a lot simpler.

Hand-edited the FOMOD xml to add all the layouts to the list of
options as well as tidy up the file structures. This felt faster
than using any gui tools to edit it.

I have not yet tested the FOMOD so more changes are no doubt to
come.