fix #94 Dropped/Sold items can't be removed from HUD (#95)
tl;dr copy-paste considered harmful. Dijkstra is alleged to have
provided that title for the essay, I have read. He was a real
trickster figure.

Plus, some efficiency tweaks for this code path. Added some
comments explaining the problem with knowing when to advance
a cycle in the drop/sell case.