Dry-coded pass one at the fix.
This filters through extra data at the time an item is equipped
and passes the specific set of relevant extra data through to the
equip call. As the comments note, we are not yet matching for
specific names so we get the enchanted version of an item instead
of all the generic versions. (Bleah.)