Really quite a lot of namespace thrash, and one useful new feature.
There's a new struct for all the extra data that hud items care
about, and some plumbing so HudItem can ask the game to fill out
the struct. Only the `is-enchanted` extra data is immutable,
really, so we can hit this on a timer or something and update
objects when we think we need to with one call.