Extra data is now refreshed every N draw calls
Where N is a constant set in ui_renderer.cpp. Removed some things we
are unlikely to call on their own, now that we're updating all extra
data on a schedule. This is currently set to 50 draw calls, which is
plenty snappy. I should experiment with making that longer.

Also, figured out how torch burn time works, so we can display the
time as a percentage left. It's kinda cool to watch the time tick
down.

Ran a format pass on the Rust files.