Rework how we cache extra data.
Get rid of the fields on HudItem and use the extra data struct
directly. Expand the struct to hold the underlying data used
to do the percentage calculations, and make all that data
available as format vars for text fields.

The max shout cooldown time must still live on the hud item,
because we still have to snag the largest value we see there and
hold onto it separately.

Updated the docs to match.