Integrate with moreHUD Inventory Edition using the iEquip hooks
This works, almost, but I'm setting the branch aside for reasons
detailed at the bottom of this commit message.

This uses SKSE ModCallbackEvents to signal papyrus from C++. The
papyrus uses the functions in MoreHUDIE to do what iEquip does and set
decorators on inventory items.

There's now a `controller.set_decorator()` function, called after all
cycle editing events to assign the appropriate decorator. We now call
it after favoriting and after toggling in menus. We also call it at
initialization time, after cycle validation.

The script is attached to the same player ref alias used by the MCM
script, which is on the MCM config quest. Quests and the story manager
do a lot of work in the Creation Engine, I guess. The script fires up
on the event onPlayerLoadGame() and registers for the mod events
mentioned above.

The moreHUD flash file `baseIcons.swf` has changes to add the new
icons needed by Soulsy. These are a left indicator, a right indicator,
a both-hands indicator and an asterisk used by shouts & utility items.

There are several problems with this work that result in me setting it
aside for the moment. First, I had to make changes to the core
moreHUDIE scripts to suppress the name of the indicator showing up in
text on potion item cards. Second, the shouts item cards show the
indicator above the card, and I haven't figured this out yet. Third,
only one weapon gets an indicator at any time. This is likely fixable,
but would require more hacking on the core of moreHUDIE.

Serialize the changes in the plugin, for later study.