feat(layouts): override layout anchor points with user settings (#108)
Players can now choose to override a layout's chosen anchor point
with their own preferred HUD location. This won't work well for
every layout, but players can decide that for themselves.

Added the new option to the settings struct and MCM config. It's an 
enum that allows players to choose from the current list of anchor 
names. Added new translation strings to support the config.

`apply_settings()` now refreshes the layout to pick up any
location changes.

Added a test to ensure that the setting is respected by the
anchor_point() function if set. This bottleneck ensures that
flattened layouts get the overridden location automatically.
To support the test, added a way to specific a refresh-from file
for the settings lazy static. The new convenience functions to
Layout that match on its variants are also there to support tests.