fix: Lock the HUD down when in vampire lord form (#109)
This fixes #100 : handle being a vampire lord more gracefully.

Added two new checks to the player conveniences: `isVampireLord()` and 
`isWerewolf()`. These check the player's race and return true if appropriate. 
Put the vampire lord check to work in deciding if we should respond to 
keystrokes at all. This effectively locks the HUD out. The HUD cannot do 
anything useful for vampire lords, because they are restricted to a tiny handful 
of spells and powers, all of which must be selected from the favorites menu.

While testing this work, observed some things with vampire spells that could 
be better. This resulted in some new icons, some color selection reworks, 
and a bit of code cleanup.