feat(ux): I4 icons as an optional indicator in menus
The splat indicates presence in cycle where there is no ambiguity
about which cycle is meant. `<-` left arrow, `->` right arrow and
`<.>` left/right arrows are used for weapons and spells or anything
else that could be in one or both of the hand cycles.

The controller now sets the correct keywords for items in cycles
at load time as well as when items are added or removed.

Remaining tasks:
- add keywords to the ESP
- test in-game
- add a setting to make this optional

Added arrow svg assets in case later tweaking becomes necessary.