Apply the scaling factor now that we know it.
Scale everything down if we're downscaling, and twice if
we're upscaling after the downscale. Hand-specified anchor
points are not affected by this scaling work, because we
assume people know what they're doing. (Possibly a bad
assumption.) An upscaled layout doesn't look good, but at
least it's now a good size and in the right place.