feature: Let's start scaling when display tweaks is active (#118)
We now sneak a peek at the DisplayTweaks settings if we can find them,
and adjust our sizes to account for any scaling or upscaling done
behind Imgui's back.

If the window is downscaled, we include that in the layout's global
scaling when flattening. If it's being upscaled afterward, we
downscale twice to stay at a reasonable size. I do not think upscaling
looks good, but at least everything is in the right place now.

Anchors specified via x & y coordinates are not affected by this
scaling.