feat(layouts): Text elements can now be truncated
If you add set truncate to true in a text element, like this:

```toml
wrap_width = 130.0
truncate = true
```

The text is truncated at the given wrap width instead of wrapped.
Strictly speaking we should change the field `wrap_width` to
`width`, but I fear breaking existing layouts. Dry-coded, so it's
not going to work, but hey. Why change my habits now?

This should fix #104 when it lands.