Lines are broken on spaces then aligned one by one.
Left-aligned lines go through the built-in imgui
line-breaking algorithm. For right- and center-aligned
lines, we let imgui tell us where the line would break
to hit max width, then walk left a few characters to find
the nearest space if one exists. If we find one, we use
that to break and recalculate our bounds. If we don't,
we use what imgui suggested.

This looks good for all of the cases, I believe.

Hooray!