fix(hud): HUD not hidden when using Skyrim Souls
Any mod that un-pauses when menus are shown would trigger this
bug. The fix is to track the visibility of the *cursor* via the
cursor menu. If this is visible, the HUD should not be visible.

There are other cases where we should hide ourselves, but we
already handle those, and those checks are untouched here.
(It turned out pretty simple in the end!)