fix: handle Skyrim Souls even better
We now pay attention to two different lists of menus:
menus that should force the HUD to hide itself, and menus
that should force the HUD to ignore keypresses. We used to
rely on the game paused status to make this decision, which
of course the souls-style menu unpausers make impossible.

I've played for a while with unpaused menus and it looks
good so far.